requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
pillow==10.4.0
djangorestframework-simplejwt==5.3.1
filelock==3.15.1
idna==3.7
platformdirs==4.2.2
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.2
virtualenvwrapper-win==1.2.7
